moves page tests out to their own folder
