+++ next.config.js
@@ -3,10 +3,10 @@ module.exports = {
 		ignoreDuringBuilds: true,
 	},
     exportPathMap: () => ({
-        // '/': { page: '/' },
-        // '/guide': { page: '/guide' },
-        // '/shoppingList': { page: '/shoppingList' },
-        // ...recipePaths,
+        '/': { page: '/' },
+        '/guide': { page: '/guide' },
+        '/shoppingList': { page: '/shoppingList' },
     }),
+    output: 'export',
     reactStrictMode: true,
 }
