+++ src/pageTests/index.test.tsx
@@ -1,11 +1,11 @@
 import { render } from '@testing-library/react'
 import React from 'react';
 
-import GuidePage from './guide';
+import HomePage from '../pages/index';
 
-describe('Page - Guide', () => {
+describe('Page - Index', () => {
     it('should render', () => {
-        const { container } = render(<GuidePage />);
+        const { container } = render(<HomePage />);
 
         expect(container).toBeInTheDocument();
     });
